# internal/demos/demo1x/demo1go/api/v1/greeter/greeter.pb.go
// Code generated by protoc-gen-go. DO NOT EDIT.
// versions:
// 	protoc-gen-go v1.36.7
// 	protoc        v5.29.3
// source: v1/greeter/greeter.proto

package greeter

import (
	_ "google.golang.org/genproto/googleapis/api/annotations"
	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
	_ "google.golang.org/protobuf/types/descriptorpb"
	reflect "reflect"
	sync "sync"
	unsafe "unsafe"
)

const (
	// Verify that this generated code is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(20 - protoimpl.MinVersion)
	// Verify that runtime/protoimpl is sufficiently up-to-date.
	_ = protoimpl.EnforceVersion(protoimpl.MaxVersion - 20)
)

// The request message containing the user's name.
type HelloRequest struct {
	state         protoimpl.MessageState `protogen:"open.v1"`
	Name          string                 `protobuf:"bytes,1,opt,name=name,proto3" json:"name,omitempty"`
	unknownFields protoimpl.UnknownFields
	sizeCache     protoimpl.SizeCache
}

func (x *HelloRequest) Reset() {
	*x = HelloRequest{}
	mi := &file_v1_greeter_greeter_proto_msgTypes[0]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *HelloRequest) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*HelloRequest) ProtoMessage() {}

func (x *HelloRequest) ProtoReflect() protoreflect.Message {
	mi := &file_v1_greeter_greeter_proto_msgTypes[0]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use HelloRequest.ProtoReflect.Descriptor instead.
func (*HelloRequest) Descriptor() ([]byte, []int) {
	return file_v1_greeter_greeter_proto_rawDescGZIP(), []int{0}
}

func (x *HelloRequest) GetName() string {
	if x != nil {
		return x.Name
	}
	return ""
}

// The response message containing the greetings
type HelloReply struct {
	state         protoimpl.MessageState `protogen:"open.v1"`
	Message       string                 `protobuf:"bytes,1,opt,name=message,proto3" json:"message,omitempty"`
	unknownFields protoimpl.UnknownFields
	sizeCache     protoimpl.SizeCache
}

func (x *HelloReply) Reset() {
	*x = HelloReply{}
	mi := &file_v1_greeter_greeter_proto_msgTypes[1]
	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
	ms.StoreMessageInfo(mi)
}

func (x *HelloReply) String() string {
	return protoimpl.X.MessageStringOf(x)
}

func (*HelloReply) ProtoMessage() {}

func (x *HelloReply) ProtoReflect() protoreflect.Message {
	mi := &file_v1_greeter_greeter_proto_msgTypes[1]
	if x != nil {
		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
		if ms.LoadMessageInfo() == nil {
			ms.StoreMessageInfo(mi)
		}
		return ms
	}
	return mi.MessageOf(x)
}

// Deprecated: Use HelloReply.ProtoReflect.Descriptor instead.
func (*HelloReply) Descriptor() ([]byte, []int) {
	return file_v1_greeter_greeter_proto_rawDescGZIP(), []int{1}
}

func (x *HelloReply) GetMessage() string {
	if x != nil {
		return x.Message
	}
	return ""
}

var File_v1_greeter_greeter_proto protoreflect.FileDescriptor

const file_v1_greeter_greeter_proto_rawDesc = "" +
	"\n" +
	"\x18v1/greeter/greeter.proto\x12\x0eapi.v1.greeter\x1a\x1cgoogle/api/annotations.proto\x1a google/protobuf/descriptor.proto\"\"\n" +
	"\fHelloRequest\x12\x12\n" +
	"\x04name\x18\x01 \x01(\tR\x04name\"&\n" +
	"\n" +
	"HelloReply\x12\x18\n" +
	"\amessage\x18\x01 \x01(\tR\amessage2q\n" +
	"\aGreeter\x12f\n" +
	"\bSayHello\x12\x1c.api.v1.greeter.HelloRequest\x1a\x1a.api.v1.greeter.HelloReply\" \x82\xd3\xe4\x93\x02\x1a\x12\x18/api/greeter/meet/{name}B2\n" +
	"\x0eapi.v1.greeterP\x01Z\x1edemo1go/api/v1/greeter;greeterb\x06proto3"

var (
	file_v1_greeter_greeter_proto_rawDescOnce sync.Once
	file_v1_greeter_greeter_proto_rawDescData []byte
)

func file_v1_greeter_greeter_proto_rawDescGZIP() []byte {
	file_v1_greeter_greeter_proto_rawDescOnce.Do(func() {
		file_v1_greeter_greeter_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_v1_greeter_greeter_proto_rawDesc), len(file_v1_greeter_greeter_proto_rawDesc)))
	})
	return file_v1_greeter_greeter_proto_rawDescData
}

var file_v1_greeter_greeter_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
var file_v1_greeter_greeter_proto_goTypes = []any{
	(*HelloRequest)(nil), // 0: api.v1.greeter.HelloRequest
	(*HelloReply)(nil),   // 1: api.v1.greeter.HelloReply
}
var file_v1_greeter_greeter_proto_depIdxs = []int32{
	0, // 0: api.v1.greeter.Greeter.SayHello:input_type -> api.v1.greeter.HelloRequest
	1, // 1: api.v1.greeter.Greeter.SayHello:output_type -> api.v1.greeter.HelloReply
	1, // [1:2] is the sub-list for method output_type
	0, // [0:1] is the sub-list for method input_type
	0, // [0:0] is the sub-list for extension type_name
	0, // [0:0] is the sub-list for extension extendee
	0, // [0:0] is the sub-list for field type_name
}

func init() { file_v1_greeter_greeter_proto_init() }
func file_v1_greeter_greeter_proto_init() {
	if File_v1_greeter_greeter_proto != nil {
		return
	}
	type x struct{}
	out := protoimpl.TypeBuilder{
		File: protoimpl.DescBuilder{
			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
			RawDescriptor: unsafe.Slice(unsafe.StringData(file_v1_greeter_greeter_proto_rawDesc), len(file_v1_greeter_greeter_proto_rawDesc)),
			NumEnums:      0,
			NumMessages:   2,
			NumExtensions: 0,
			NumServices:   1,
		},
		GoTypes:           file_v1_greeter_greeter_proto_goTypes,
		DependencyIndexes: file_v1_greeter_greeter_proto_depIdxs,
		MessageInfos:      file_v1_greeter_greeter_proto_msgTypes,
	}.Build()
	File_v1_greeter_greeter_proto = out.File
	file_v1_greeter_greeter_proto_goTypes = nil
	file_v1_greeter_greeter_proto_depIdxs = nil
}